Moved logical function to a new file in the analyzer module called abilityChunksCalc.ts - Refactored for loop content to be cleaner - Removed & changed some comments
