+++ app/routes/analyzer.tsx
@@ -19,7 +19,6 @@ import {
 } from "~/modules/analyzer/specialEffects";
 import type {
   AbilityPoints,
-  MainAbilityChunks,
   SpecialEffectType,
 } from "~/modules/analyzer/types";
 import {
@@ -48,10 +47,7 @@ import {
   subWeaponImageUrl,
 } from "~/utils/urls";
 import type { AbilityWithUnknown } from "~/modules/in-game-lists/types";
-import {
-  PRIMARY_SLOT_ONLY_REQUIRED_ABILITY_CHUNKS_COUNT,
-  REQUIRED_ABILITY_CHUNKS_COUNT,
-} from "~/modules/in-game-lists/abilities";
+import { getAbilityChunksMapAsArray } from "~/modules/analyzer/abilityChunksCalc";
 
 export const CURRENT_PATCH = "1.2";
 
@@ -749,57 +745,6 @@ function AbilityPointsDetails({
       </div>
     </details>
   );
-}
-
-/**
- * From an array of Main abilities, create a map of <Ability, number>, then return it as an Array after sorting by value, descending.
- *
- * The map describes the number of Ability chunks required to replace the main ability on a piece of gear.
- *
- * Extra processing is required for Main abilities that are primary slot-only abilities,
- *  as they are comprised of 3 stackable ability chunks at a lower ability chunk count than usual.
- */
-function getAbilityChunksMapAsArray(mainAbilities: AbilityWithUnknown[]) {
-  const abilityChunksMap: MainAbilityChunks = new Map();
-  for (let i = 0; i < mainAbilities.length; i++) {
-    const mainAbility = mainAbilities[i];
-    if (typeof mainAbility !== "undefined" && mainAbility !== "UNKNOWN") {
-      const primarySlotOnlyAbilityRef = abilities.filter(
-        (a) => a.name === mainAbility && a.abilityChunkTypesRequired.length > 0
-      );
-
-      // Primary-slot-only item that can have ability chunks put on it
-      if (primarySlotOnlyAbilityRef.length === 1) {
-        const primaryAbility = primarySlotOnlyAbilityRef[0];
-        if (typeof primaryAbility !== "undefined") {
-          for (
-            let j = 0;
-            j < primaryAbility?.abilityChunkTypesRequired.length;
-            j++
-          ) {
-            const mainAbilityName = primaryAbility.abilityChunkTypesRequired[j];
-
-            if (typeof mainAbilityName !== "undefined") {
-              abilityChunksMap.set(
-                mainAbilityName,
-                (abilityChunksMap.get(mainAbility) ?? 0) +
-                  PRIMARY_SLOT_ONLY_REQUIRED_ABILITY_CHUNKS_COUNT
-              );
-            }
-          }
-        }
-      } else {
-        abilityChunksMap.set(
-          mainAbility,
-          (abilityChunksMap.get(mainAbility) ?? 0) +
-            REQUIRED_ABILITY_CHUNKS_COUNT
-        );
-      }
-    }
-  }
-
-  // After building the map, convert it to an array & sort the map by value, descending
-  return Array.from(abilityChunksMap).sort((a, b) => b[1] - a[1]);
 }
 
 function AbilityChunksRequired({
